Extract PropertyHooksProcessor from NodeScopeResolver
